Penerapan middleware pada routes.

+++ routes/api.php
@@ -24,18 +24,18 @@ Route::post('/reset-password', [PasswordController::class, 'resetPassword']);
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     Route::post('/logout', [AuthenticationController::class, 'logout']);
     Route::get('/me', [AuthenticationController::class, 'me']);
-    Route::post('/change-password', [PasswordController::class, 'changePassword']);
+    Route::post('/change-password', [PasswordController::class, 'changePassword'])->middleware('account-owner');
 
-    Route::get('/users/admins', [UserController::class, 'indexAdmins']);
-    Route::get('/users/admins/{id}', [UserController::class, 'showAdmin']);
+    Route::get('/users/admins', [UserController::class, 'indexAdmins'])->middleware('admin-only');
+    Route::get('/users/admins/{id}', [UserController::class, 'showAdmin'])->middleware('admin-only');
 
     Route::get('/articles', [ArticleController::class, 'index']);
     Route::get('/articles/{id}', [ArticleController::class, 'show']);
-    Route::post('/articles', [ArticleController::class, 'store']);
-    Route::patch('/articles/{id}', [ArticleController::class, 'update']);
-    Route::delete('/articles/{id}', [ArticleController::class, 'destroy']);
+    Route::post('/articles', [ArticleController::class, 'store'])->middleware('admin-or-doctor');
+    Route::patch('/articles/{id}', [ArticleController::class, 'update'])->middleware('article-owner');
+    Route::delete('/articles/{id}', [ArticleController::class, 'destroy'])->middleware('article-owner');
     
     Route::post('/comments', [CommentController::class, 'store']);
-    Route::patch('/comments/{id}', [CommentController::class, 'update']);
-    Route::delete('/comments/{id}', [CommentController::class, 'destroy']);
+    Route::patch('/comments/{id}', [CommentController::class, 'update'])->middleware('comment-owner');
+    Route::delete('/comments/{id}', [CommentController::class, 'destroy'])->middleware('comment-owner');
 });
